refactor(Player): Participant에서 Player로 이름 변경
