+++ src/main/java/blackjack/player/state/terminated/BustedState.java
@@ -1,0 +1,17 @@
+package blackjack.player.state.terminated;
+
+import blackjack.game.MatchResult;
+import blackjack.player.Hand;
+import blackjack.player.state.GameState;
+
+public class BustedState extends TerminatedState {
+
+    public BustedState(Hand hand) {
+        super(hand);
+    }
+
+    @Override
+    public MatchResult createMatchResult(GameState opponentState) {
+        return MatchResult.LOSE;
+    }
+}
