+++ src/main/java/blackjack/player/state/terminated/TerminatedState.java
@@ -1,0 +1,27 @@
+package blackjack.player.state.terminated;
+
+import blackjack.card.Card;
+import blackjack.player.Hand;
+import blackjack.player.state.GameState;
+
+public abstract class TerminatedState extends GameState {
+
+    protected TerminatedState(Hand hand) {
+        super(hand);
+    }
+
+    @Override
+    public GameState drawCard(Card card) {
+        throw new UnsupportedOperationException("[ERROR] 현재 상태에서는 드로우 여부를 결정할 수 없습니다.");
+    }
+
+    @Override
+    public GameState stand() {
+        throw new UnsupportedOperationException("[ERROR] 현재 상태에서는 스탠드 여부를 결정할 수 없습니다.");
+    }
+
+    @Override
+    public boolean isTerminated() {
+        return true;
+    }
+}
